+++ nix/chainmain.nix
@@ -1,6 +1,3 @@
-{ pkgs ? import ./default.nix { } }:
-(import
-  (builtins.fetchTarball
-    "https://github.com/crypto-org-chain/chain-main/archive/v2.1.2.tar.gz")
-  { }).chain-maind
+{ pkgs ? import ./default.nix { }, sources ? import ./sources.nix }:
+(import sources.chain-main { }).chain-maind