+++ x/icactl/keeper/grpc_query.go
@@ -1,0 +1,7 @@
+package keeper
+
+import (
+	"github.com/crypto-org-chain/cronos/x/icactl/types"
+)
+
+var _ types.QueryServer = Keeper{}